Change update repair request to follow rest api convention
